Revert "Revert "clean up of unnecessary either T methods, small changes to im… (#137)"
This reverts commit 89902b486164f849f0cf24467a4a7f0a6c0e36f2.
